+++ test/e2e/framework/BUILD
@@ -20,6 +20,7 @@ go_library(
         "rc_util.go",
         "resource_usage_gatherer.go",
         "size.go",
+        "skip.go",
         "suites.go",
         "test_context.go",
         "util.go",
